+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+### Changed
+
+- Fix Table.Row detection in [Table](/#/Components?id=table).
+
 ## [6.0.0] - 2019-05-07
 
 ### Changed